+++ blobstore_client/lib/blobstore_client/local_client.rb
@@ -12,9 +12,12 @@ module Bosh
         FileUtils.mkdir_p(@blobstore_path) unless File.directory?(@blobstore_path)
       end
 
-      def create_file(file)
-        id = UUIDTools::UUID.random_create.to_s
-        dst = File.join(@blobstore_path, id)
+      protected
+
+      def create_file(id, file)
+        id ||= UUIDTools::UUID.random_create.to_s
+        dst = object_file_path(id)
+        raise BlobstoreError, "object id #{id} is already in use" if File.exist?(dst)
         File.open(dst, 'w') do |fh|
           until file.eof?
             fh.write(file.read(CHUNK_SIZE))
@@ -24,7 +27,7 @@ module Bosh
       end
 
       def get_file(id, file)
-        src = File.join(@blobstore_path, id)
+        src = object_file_path(id)
 
         begin
           File.open(src, 'r') do |src_fh|
@@ -37,11 +40,21 @@ module Bosh
         raise NotFound, "Blobstore object '#{id}' not found"
       end
 
-      def delete(id)
-        file = File.join(@blobstore_path, id)
+      def delete_object(id)
+        file = object_file_path(id)
         FileUtils.rm(file)
       rescue Errno::ENOENT
         raise NotFound, "Blobstore object '#{id}' not found"
+      end
+
+      def object_exists?(oid)
+        File.exists?(object_file_path(oid))
+      end
+
+      private
+
+      def object_file_path(oid)
+        File.join(@blobstore_path, oid)
       end
 
     end
